+++ CHANGES.rst
@@ -43,6 +43,8 @@ Bugfixes
 - No longer silently fall back to the first event contact email address when sending
   registration emails where no explicit sender address has been configured (:issue:`4992`,
   :pr:`4996`, thanks :user:`vasantvohra`)
+- Do not check for event access when using a registration link with a registration token
+  (:issue:`4991`, :pr:`4997`, thanks :user:`vasantvohra`)
 
 Version 3.0rc1
 --------------